Reorganise new classes
Reorganise MainWindow and DialogBox into gui package.

Reorganise Main and Launcher into main package.
