+++ src/main/java/rakan/main/Launcher.java
@@ -1,4 +1,4 @@
-package rakan;
+package rakan.main;
 
 import javafx.application.Application;
